+++ src/encoding/typing.py
@@ -1,0 +1,7 @@
+from typing import NewType
+
+
+U256 = NewType("U256", int)
+U8 = NewType("U8", int)
+# must be one of -1, 0, 1
+Sign = NewType("Sign", int)
